Peptide position starts at 1 for first residue for all reporting.
Currently, MetaMorpheus reports dynamic and static mods starting at 2
for the first residue. And it starts at 1 for deadend, cross-link, and
loop-link locations. The developers of MetaMorpheus are going to
standardize on reporting such that 1 is the first residue of the peptide
for all positional reporting.

I have preemptively made the change in this converter for that upcoming
change. After version 277 of MetaMorpheus, locations will be correctly
reported in proxl XML.